# docs/swagger.yaml
basePath: /
definitions:
  dto.AttachmentResponse:
    properties:
      file_type:
        description: image, pdf, etc.
        type: string
      url:
        type: string
    type: object
...
  dto.CreatePostRequest:
    properties:
      content:
        minLength: 1
        type: string
      parent_id:
        description: если это комментарий
        type: string
    required:
    - content
    type: object
...
  dto.PostLiteResponse:
    properties:
      attachments:
        items:
          $ref: '#/definitions/dto.AttachmentResponse'
        type: array
      author_id:
        type: string
      avatar_url:
        type: string
      comments:
        type: integer
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      likes:
        type: integer
      microtask_id:
        type: string
      microtask_title:
        type: string
      promise_id:
        type: string
      promise_title:
        type: string
      username:
        type: string
    type: object
  dto.PostWithRepliesTreeResponse:
    properties:
      attachments:
        items:
          $ref: '#/definitions/dto.AttachmentResponse'
        type: array
      author_id:
        type: string
      avatar_url:
        type: string
      comments:
        type: integer
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      likes:
        type: integer
      microtask_id:
        type: string
      microtask_title:
        type: string
      promise_id:
        type: string
      promise_title:
        type: string
      replies:
        items:
          $ref: '#/definitions/dto.PostWithRepliesTreeResponse'
        type: array
      username:
        type: string
    type: object
...
  dto.UpdatePostRequest:
    properties:
      content:
        type: string
    required:
    - content
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API для отслеживания обещаний пользователей.
  title: iPromise API
  version: "1.0"
paths:
...
  /microtasks/{microtask_id}/posts:
    get:
      description: Возвращает все корневые посты, связанные с микротаском
      parameters:
      - description: ID микротаска
        in: path
        name: microtask_id
        required: true
        type: string
      - description: Максимальное количество
        in: query
        name: limit
        type: integer
      - description: Дата (RFC3339) для пагинации
        in: query
        name: after
        type: string
      - description: ID поста для пагинации
        in: query
        name: after_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.PostWithRepliesTreeResponse'
            type: array
        "400":
          description: 'error: Неверный ID'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Неавторизован'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: Нет доступа'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Ошибка сервера'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получить посты по микротаску
      tags:
      - posts
    post:
      consumes:
      - application/json
      description: Создаёт корневой пост или комментарий (если указан parent_id)
      parameters:
      - description: ID микротаска
        in: path
        name: microtask_id
        required: true
        type: string
      - description: Данные поста
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePostRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 'message: Пост создан'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'error: Неверный формат'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Неавторизован'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: Нет прав доступа'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Ошибка сервера'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Создать пост или комментарий
      tags:
      - posts
  /posts/{id}:
    delete:
      description: Удаляет пост по ID (soft delete)
      parameters:
      - description: ID поста
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Пост удалён'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'error: Неверный ID'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Неавторизован'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: Нет прав'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Ошибка сервера'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удалить пост
      tags:
      - posts
    patch:
      consumes:
      - application/json
      description: Обновляет содержимое поста
      parameters:
      - description: ID поста
        in: path
        name: id
        required: true
        type: string
      - description: Новое содержимое
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Пост обновлён'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'error: Неверный формат'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Неавторизован'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: Нет прав'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Ошибка сервера'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Обновить пост
      tags:
      - posts
  /posts/{id}/full:
    get:
      description: Возвращает пост, все его реплаи, автора, вложения и лайки
      parameters:
      - description: ID поста
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PostWithRepliesTreeResponse'
        "400":
          description: 'error: Неверный ID'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Пост не найден'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Ошибка сервера'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получить пост с деревом комментариев
      tags:
      - posts
  /posts/{id}/replies:
    get:
      description: Возвращает список дочерних постов (реплаев)
      parameters:
      - description: ID родительского поста
        in: path
        name: id
        required: true
        type: string
      - description: Максимальное количество
        in: query
        name: limit
        type: integer
      - description: Дата (RFC3339) для пагинации
        in: query
        name: after
        type: string
      - description: ID поста для пагинации
        in: query
        name: after_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.PostWithRepliesTreeResponse'
            type: array
        "400":
          description: 'error: Неверный ID'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Неавторизован'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Ошибка сервера'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получить комментарии к посту
      tags:
      - posts
  /posts/feed:
    get:
      description: Посты от тех, на кого подписан пользователь
      parameters:
      - description: Максимум постов
        in: query
        name: limit
        type: integer
      - description: Дата (RFC3339)
        in: query
        name: after
        type: string
      - description: ID последнего поста
        in: query
        name: after_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.PostLiteResponse'
            type: array
        "401":
          description: 'error: Неавторизован'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Ошибка сервера'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Лента подписок (лайт)
      tags:
      - posts
  /posts/feed/tree:
    get:
      description: Посты с реплаями от пользователей, на которых подписан текущий
      parameters:
      - description: Максимум постов
        in: query
        name: limit
        type: integer
      - description: Дата (RFC3339)
        in: query
        name: after
        type: string
      - description: ID последнего поста
        in: query
        name: after_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.PostWithRepliesTreeResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Лента подписок (с деревом)
      tags:
      - posts
  /posts/public:
    get:
      description: Публичная лента постов без дерева комментариев
      parameters:
      - description: Максимум постов
        in: query
        name: limit
        type: integer
      - description: Дата (RFC3339) для пагинации
        in: query
        name: after
        type: string
      - description: ID последнего поста
        in: query
        name: after_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.PostLiteResponse'
            type: array
        "401":
          description: 'error: Неавторизован'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Ошибка сервера'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Публичные посты (лайт)
      tags:
      - posts
  /posts/public/tree:
    get:
      description: Публичная лента постов с комментариями
      parameters:
      - description: Максимум постов
        in: query
        name: limit
        type: integer
      - description: Дата (RFC3339)
        in: query
        name: after
        type: string
      - description: ID последнего поста
        in: query
        name: after_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.PostWithRepliesTreeResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Публичные посты (с деревом)
      tags:
      - posts
  /posts/user/{username}:
    get:
      description: Получает все посты пользователя по username, включая комментарии
        (если приватные – только для владельца)
      parameters:
      - description: Имя пользователя (username)
        in: path
        name: username
        required: true
        type: string
      - description: Максимум постов
        in: query
        name: limit
        type: integer
      - description: Дата (RFC3339) для пагинации
        in: query
        name: after
        type: string
      - description: ID последнего поста
        in: query
        name: after_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.PostWithRepliesTreeResponse'
            type: array
        "400":
          description: 'error: Некорректный username'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Неавторизован'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Ошибка сервера'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Посты пользователя (с деревом)
      tags:
      - posts
...
  /promises/{promise_id}/posts:
    get:
      description: Возвращает все посты, относящиеся к данному промису
      parameters:
      - description: ID промиса
        in: path
        name: promise_id
        required: true
        type: string
      - description: Максимум постов
        in: query
        name: limit
        type: integer
      - description: Дата (RFC3339) пагинации
        in: query
        name: after
        type: string
      - description: ID последнего поста
        in: query
        name: after_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.PostWithRepliesTreeResponse'
            type: array
        "400":
          description: 'error: Неверный ID'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 'error: Неавторизован'
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 'error: Нет доступа'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'error: Ошибка сервера'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получить посты по промису
      tags:
      - posts
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"